chore: rename component files to [ComponentName].tsx
